+++ src/components/card-screen/card-screen.jsx
@@ -21,7 +21,7 @@ import Footer from '../footer/footer.jsx';
 import compose from '../../hocs/compose/compose.js';
 import withCardTabsState from '../../hocs/with-card-tabs-state/with-card-tabs-state.jsx';
 import withPreviewCardListState from '../../hocs/with-preview-card-list-state/with-preview-card-list-state.jsx';
-import withData from '../../hocs/with-data/with-data.jsx';
+import withFetchData from '../../hocs/with-fetch-data/with-fetch-data.jsx';
 import withActiveItem from '../../hocs/with-active-item/with-active-item.jsx';
 import {getCardsData} from '../../reducers/card-list/selectors.js';
 
@@ -30,7 +30,7 @@ import {DataTypes, ComponentTypes} from '../../const.js';
 
 const WrappedCardTabs = withCardTabsState(CardTabs);
 
-const WrappedCardReviews = withData(DataTypes.FETCH_REVIEWS_DATA)(CardReviews);
+const WrappedCardReviews = withFetchData(DataTypes.FETCH_REVIEWS_DATA)(CardReviews);
 
 const WrappedPreviewCardList = compose(
     withActiveItem(ComponentTypes.PREVIEW_CARDS_LIST),
