+++ src/components/main-content/main-content.test.js
@@ -16,7 +16,7 @@ import Footer from '../footer/footer.jsx';
 import Logo from '../logo/logo.jsx';
 
 import withPreviewCardListState from '../../hocs/with-preview-card-list-state/with-preview-card-list-state.jsx';
-import withData from '../../hocs/with-data/with-data.jsx';
+import withFetchData from '../../hocs/with-fetch-data/with-fetch-data.jsx';
 import withActiveItem from '../../hocs/with-active-item/with-active-item.jsx';
 import compose from '../../hocs/compose/compose.js';
 
@@ -103,7 +103,7 @@ const WrappedPreviewCardList = compose(
     withPreviewCardListState)(PreviewCardList);
 
 const WrappedGenreList = compose(
-    withData(DataTypes.FETCH_CARDS_DATA),
+    withFetchData(DataTypes.FETCH_CARDS_DATA),
     withActiveItem(ComponentTypes.GENRES_LIST))(GenreList);
