+++ src/components/main/main.jsx
@@ -9,17 +9,17 @@ import Footer from '../footer/footer.jsx';
 import Logo from '../logo/logo.jsx';
 
 import compose from '../../hocs/compose/compose.js';
-import withData from '../../hocs/with-data/with-data.jsx';
+import withFetchData from '../../hocs/with-fetch-data/with-fetch-data.jsx';
 import withActiveItem from '../../hocs/with-active-item/with-active-item.jsx';
 import withPreviewCardListState from '../../hocs/with-preview-card-list-state/with-preview-card-list-state.jsx';
 
 import {DataTypes, ComponentTypes} from '../../const.js';
 
 
-const WrappedMainHeader = withData(DataTypes.FETCH_PROMO_DATA)(MainHeader);
+const WrappedMainHeader = withFetchData(DataTypes.FETCH_PROMO_DATA)(MainHeader);
 
 const WrappedGenreList = compose(
-    withData(DataTypes.FETCH_CARDS_DATA),
+    withFetchData(DataTypes.FETCH_CARDS_DATA),
     withActiveItem(ComponentTypes.GENRES_LIST))(GenreList);
 
 const WrappedPreviewCardList = compose(
